requirements.txt
google-cloud-storage>=1.20.0
python-dotenv>=0.10.0
pyyaml>=5.1